Good progress on the meters
So the meters are gonna be plain html. a lot less manual shiz like placing
text and other bullshit that svg needs to work like a div. And we can always
sprinkle in the svg icons in there and custom svg shiz later.
so yea what's great is working with percentages makes it responsive for free
instead of a bunch of javascript